% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusUK2021.R
\docType{package}
\name{censusUK2021}
\alias{censusUK2021}
\title{UK Census 2021}
\description{
`R` package that include data, boundaries and a Shiny app to visualize interactively the new preliminary results of the last 2021 UK Census.
}
\details{
As of June 2022, data only include `population` by `sex` and `age` in five-years classes plus `households`, 
with the lowest geographical detail being `Local Authority Districts` in England and Wales only.
 
Attributions:
\describe{
  \item{ Contains OS data © Crown copyright and database rights [2021] }
  \item{ Contains National Statistics data © Crown copyright and database rights [2021] }
  \item{ Source: Office for National Statistics licensed under the Open Government Licence v.3.0 }
}
}
