% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{czones}
\alias{czones}
\title{czones}
\format{
A data.table with the following columns:
\describe{
  \item{\code{ LAD }}{ ONS code for Local Authority Districts }
  \item{\code{ x_lon }}{ Longitude of the geometric centroid for Local Authority Districts }
  \item{\code{ y_lat }}{ Latitude of the geometric centroid for Local Authority Districts }
  \item{\code{ area }}{ Area for Local Authority Districts }
  \item{\code{ LADn }}{ ONS description for Local Authority Districts}
  \item{\code{ UTLA }}{ ONS code for Upper Tier Local Authority }
  \item{\code{ UTLAn }}{ ONS description for Upper Tier Local Authority }
  \item{\code{ RGN }}{ ONS code for Region (England only) }
  \item{\code{ RGNn  }}{ ONS description for Region (England only) }
  \item{\code{ CTRY  }}{ ONS code for Country }
  \item{\code{ CTRYn  }}{ ONS description for Country }
}

For more details, consult the website \url{https://geoportal.statistics.gov.uk/}
}
\usage{
czones
}
\description{
List of all zones in England and Wales included in "dataset", displayed as a lookup table
}
\keyword{datasets}
