% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{dataset}
\format{
A data.table with the following columns:
\describe{
  \item{\code{ period }}{ Year of Census: 1981 to 2021 }
  \item{\code{ zone_id }}{ ONS code for the zone (see table `zones`) }
  \item{\code{ var_id }}{ Code for the variable (see column `id` in table `vars`) }
  \item{\code{ value }}{ Value for the variable at those zone and year  }
}

For more details, consult the website \url{https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates/bulletins/populationandhouseholdestimatesenglandandwales/census2021}
}
\usage{
dataset
}
\description{
Data for Local Authority Districts (and up) in England and Wales for the Census from 1981 to 2021,
}
\keyword{datasets}
