% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LAD}
\alias{LAD}
\title{LAD}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 331 rows and 2 columns.
}
\usage{
LAD
}
\description{
Geographic Boundaries in \code{sf} format for Local Authority Districts in England and Wales (last update: May 2021).
}
\details{
More information can be found in the table \code{lookups}.

For more details, consult the website \url{https://geoportal.statistics.gov.uk/}
}
\keyword{datasets}
