% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookups}
\alias{lookups}
\title{lookups}
\format{
A data.table with the following columns:
\describe{
  \item{\code{ LAD }}{ ONS code for Local Authority Districts }
  \item{\code{ x_lon }}{ Longitude of the geometric centroid for Local Authority Districts }
  \item{\code{ y_lat }}{ Latitude of the geometric centroid for Local Authority Districts }
  \item{\code{ area }}{ Area for Local Authority Districts }
  \item{\code{ LADn }}{ ONS description for Local Authority Districts}
  \item{\code{ RGN }}{ ONS code for Region (England only) }
  \item{\code{ RGNn  }}{ ONS description for Region (England only) }
  \item{\code{ CTRY  }}{ ONS code for Country }
  \item{\code{ CTRYn  }}{ ONS description for Country }
}
}
\usage{
lookups
}
\description{
List of all Local Authority Districts in England and wales, with correspondent Region
}
\keyword{datasets}
