% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RGN}
\alias{RGN}
\title{RGN}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 10 rows and 2 columns.
}
\usage{
RGN
}
\description{
Geographic Boundaries in \code{sf} format for Regions in England (last update: May 2021).
}
\details{
More information can be found in the table \code{lookups}.

For more details, consult the website \url{https://geoportal.statistics.gov.uk/}
}
\keyword{datasets}
