% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tables}
\alias{tables}
\title{tables}
\format{
A data.table with the following columns:
\describe{
  \item{\code{ id }}{ ONS code for the table }
  \item{\code{ description }}{ Description of the content of the table }
  \item{\code{ lowest_geo }}{ Minimum geographic level for the data (see table `zones`) }
  \item{\code{ countries }}{ Countries to which data for each table are related }
}

For more details, consult the website \url{https://www.ons.gov.uk/releases/initialfindingsfromthe2021censusinenglandandwales}
}
\usage{
tables
}
\description{
List of the tables disseminated with preliminary results from ONS for Census 2021 (28-Jun-2022)
}
\keyword{datasets}
