% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UTLA}
\alias{UTLA}
\title{UTLA}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 113 rows and 2 columns.
}
\usage{
UTLA
}
\description{
Geographic Boundaries in \code{sf} format for Upper Tier Local Authority in England and Wales (last update: Dec 2021).
}
\details{
More information can be found in the table \code{lookups}.

For more details, consult the website \url{https://geoportal.statistics.gov.uk/}
}
\keyword{datasets}
