% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vars}
\alias{vars}
\title{vars}
\format{
A data.table with the following columns:
\describe{
  \item{\code{ id }}{ Code for the variable }
  \item{\code{ table_id }}{ Identifier for the table the var is included into (see column `id` in table `table`) }
  \item{\code{ description }}{ Long Description for the variable }
  \item{\code{ slabel }}{ Short Label for Chart and Plots }
  \item{\code{ label }}{ Label for Chart and Plots }
  \item{\code{ reference }}{ Reference variable id for rates and proportions }
  \item{\code{ historic }}{ For query the correct variables depending on current only or historic trend:
                             0- current only, 
                             1- both current and historic,
                             2- historic only. 
                             To query current: <= 1, to query both: >= 1.    
   }
  \item{\code{ ordering }}{ Columns to reference for ordering the variables in the same table  }
}

For more details, consult the website \url{https://www.ons.gov.uk/releases/initialfindingsfromthe2021censusinenglandandwales}
}
\usage{
vars
}
\description{
List of the variables disseminated with preliminary results from ONS for Census 2021 (28-Jun-2022)
}
\keyword{datasets}
