% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusUK2021.R
\docType{package}
\name{censusUK2021}
\alias{censusUK2021}
\title{UK Census 2021}
\description{
$R$  package that include data, boundaries and a Shiny app to visualize interactively the new preliminary results of the last 2021 UK Census.
As of June 2022, data only include "population" by "sex" and "age" in five-years classes plus "households", 
with the lowest geographical detail being "Local Authority Districts" in England and Wales only.
}
